requirements.txt
fastapi>=0.96.0
uvicorn>=0.22.0
diffusers>=0.17.0
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.19.0
Pillow>=9.5.0
python-multipart>=0.0.6
pydantic>=1.10.8
numpy>=1.24.3
openai==0.28.0
xformers==0.0.29.post3
peft==0.14.0
sentencepiece==0.2.0
python-dotenv>=0.19.0
protobuf==5.29.3